Revert "Add Bulk Edit Variables"
This reverts commit cd1df78855b77410bb2daf6b285c06389d6e622a.
